pyproject.toml
[project]
name = "moviegpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.4",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-genai>=1.21.1",
    "mysql-connector-python>=9.3.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
]
